src/test/javascript/spec/app/entities/filelist-egagement-type/filelist-egagement-type.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { FilelistEgagementTypeService } from 'app/entities/filelist-egagement-type/filelist-egagement-type.service';
import { IFilelistEgagementType, FilelistEgagementType } from 'app/shared/model/filelist-egagement-type.model';

describe('Service Tests', () => {
  describe('FilelistEgagementType Service', () => {
    let injector: TestBed;
    let service: FilelistEgagementTypeService;
    let httpMock: HttpTestingController;
    let elemDefault: IFilelistEgagementType;
    let expectedResult: IFilelistEgagementType | IFilelistEgagementType[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(FilelistEgagementTypeService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new FilelistEgagementType(0, false);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a FilelistEgagementType', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new FilelistEgagementType()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a FilelistEgagementType', () => {
        const returnedFromService = Object.assign(
          {
            mandatory: true,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of FilelistEgagementType', () => {
        const returnedFromService = Object.assign(
          {
            mandatory: true,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a FilelistEgagementType', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
